Idの型の修正とそれに合わせたdocumentの変更
id  string -> int
discordId int -> ulong

+++ src/BudgetManagementBotSystem/Domain/Entities/User.cs
@@ -6,12 +6,12 @@ public class User
 {
     public int Id { get; private set; }
     public string Name { get; private set; }
-    public int DiscordUserId { get; private set; }
+    public ulong DiscordUserId { get; private set; }
     public int GroupId { get; private set; }
     public AccountRole Role { get; private set; }
     public bool IsActive { get; private set; }
 
-    public User(string name, int discordUserId, AccountRole role)
+    public User(string name, ulong discordUserId, AccountRole role)
     {
         Name = name;
         DiscordUserId = discordUserId;
